refactor(time): move fit_erf to mymath
