+++ inc/erf.h
@@ -1,0 +1,17 @@
+#include <math.h>
+#include <stdint.h>
+#include <stdlib.h>
+
+// NOLINTBEGIN(readability-identifier-length)
+int cmp_u64(const void* a, const void* b);
+/* Acklam inverse normal CDF approximation */
+double invnorm(double p);
+// NOLINTEND(readability-identifier-length)
+
+struct stats
+{
+    double stdev;
+    double mean;
+};
+
+struct stats fit_erf(uint64_t* samples, size_t size);
